backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
spotipy
pydantic
slowapi
python-dotenv
httpx